services/rag-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.15
langchain==0.0.350 